Fix some -Wsign-compare warnings

+++ applets/wncklet/window-list.c
@@ -865,7 +865,7 @@ gboolean window_list_applet_fill(MatePanelApplet* applet)
 	/* disable the item of system monitor, if not exists.
 	 * example, mate-system-monitor, o gnome-system-monitor */
 	char* programpath;
-	int i;
+	gsize i;
 
 	for (i = 0; i < G_N_ELEMENTS(system_monitors); i += 1)
 	{
@@ -910,7 +910,7 @@ gboolean window_list_applet_fill(MatePanelApplet* applet)
 
 static void call_system_monitor(GtkAction* action, TasklistData* tasklist)
 {
-	int i;
+	gsize i;
 
 	for (i = 0; i < G_N_ELEMENTS(system_monitors); i += 1)
 	{
